qcacmn: change restriction mask to uint32_t from uint8_t
This change is to update restriction mask to uint32_t from uint8_t
as per QCA_WLAN_VENDOR_ATTR_AVOID_FREQUENCY_IFACES_BITMASK.

CRs-Fixed: 3118988

+++ umac/regulatory/dispatcher/inc/reg_services_public_struct.h
@@ -1559,7 +1559,7 @@ struct ch_avoid_freq_type {
 struct ch_avoid_ind_type {
 	uint32_t ch_avoid_range_cnt;
 	struct ch_avoid_freq_type avoid_freq_range[CH_AVOID_MAX_RANGE];
-	uint8_t restriction_mask;
+	uint32_t restriction_mask;
 };
 
 /**
